requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-admin-persian-fonts==0.2
django-cleanup==6.0.0
django-etc==1.4.0
django-hitcount==1.3.5
django-recaptcha==3.0.0
django-utils-six==2.0
idna==3.4
jdatetime==4.1.0
oauthlib==3.2.2
persiantools==3.0.1
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.12
